Bug hotfix!

+++ Snapshot/Assets/Blog/BlogUIManager.cs
@@ -96,8 +96,13 @@ public class BlogUIManager : MonoBehaviour {
 			// Finally, move the photo from .../Resources/UploadQueue/ to .../Resources/PostedImages/
 			File.Move(pathToUploadQueue + imageName + ".metaphoto", pathToPostedPhotos + imageName + ".metaphoto");
 			File.Move(pathToUploadQueue + imageName + ".png", pathToPostedPhotos + imageName + ".png");
-			File.Move(pathToUploadQueue + imageName + ".metaphoto.meta", pathToPostedPhotos + imageName + ".metaphoto.meta");
-            File.Move(pathToUploadQueue + imageName + ".png.meta", pathToPostedPhotos + imageName + ".png.meta");
+
+            if (File.Exists(pathToUploadQueue + imageName + ".metaphoto.meta")) {
+                File.Move(pathToUploadQueue + imageName + ".metaphoto.meta", pathToPostedPhotos + imageName + ".metaphoto.meta");
+            }
+            if (File.Exists(pathToUploadQueue + imageName + ".png.meta")) {
+                File.Move(pathToUploadQueue + imageName + ".png.meta", pathToPostedPhotos + imageName + ".png.meta");
+            }
         }
 		PlayerProfile.profile.save ();
